pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "regscope"
version = "0.1.0"
description = "Production-grade CLI tool for downloading, structuring, and analyzing public comments on federal rulemakings from Regulations.gov"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "RegScope Contributors" },
]
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "bertopic>=0.16",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5",
    "datasketch>=1.6",
    "pymupdf>=1.23",
    "pytesseract>=0.3",
    "pdf2image>=1.16",
    "transformers>=4.35",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.scripts]
regscope = "regscope.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["regscope"]

[project.optional-dependencies]
llm = [
    "openai>=1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
